wheat_forecasting_app/requirements.txt
streamlit
pandas
numpy
tensorflow
scikit-learn
python-dotenv
requests
pyarrow